fix overflow and add reporter compile options

+++ src/onePlusLambdaAlgorithms/algorithmsImpl/adjustingParamsSolver.h
@@ -8,7 +8,7 @@ template <uint32_t subpopulations>
 struct AdjustingParamsSolver : public OneMaxSolver
 {
     using OneMaxSolver::OneMaxSolver;
-    uint32_t solve(AbstractOffspring &, std::shared_ptr<Reporter>) override;
+    uint64_t solve(AbstractOffspring &, std::shared_ptr<Reporter>) override;
 };
 
 #endif // ADJUSTINGPARAMSSOLVER_H
